software-map: move edge label geometry to c4-edge-label-geometry
Agent-Session: 01a067ce-34b7-7952-a023-ced44207d9ca
