+++ packages/progressive-review/app/src/software-map/SoftwareMap.test.ts
@@ -21,7 +21,6 @@ import {
   type SoftwareMapRelationshipSnapshot,
   type SoftwareMapResolvedSnapshot,
   c4EdgeEndpointBubbles,
-  c4EdgePointsFromSections,
   c4LayoutSignature,
   c4PreviousInlineLayoutForRelationships,
   c4ViewportForNodeReveal,
@@ -30,7 +29,6 @@ import {
   fitC4MapView,
   focusC4MapNode,
   focusC4MapNodeAndKeyboard,
-  positionC4EdgeLabels,
   runInlineC4Layout,
   runSerializedC4Layout,
 } from "./SoftwareMap";
@@ -1381,31 +1379,6 @@ describe("SoftwareMap inline C4 helpers", () => {
     }
   });
 
-  it("uses routed C4 edge sections without schema endpoint rewrites", () => {
-    const points = c4EdgePointsFromSections([
-      {
-        startPoint: { x: 100, y: 100 },
-        bendPoints: [
-          { x: 160, y: 100 },
-          { x: 160, y: 220 },
-        ],
-        endPoint: { x: 260, y: 220 },
-      },
-    ]);
-
-    expect(points).toEqual([
-      { x: 100, y: 100 },
-      { x: 160, y: 100 },
-      { x: 160, y: 220 },
-      { x: 260, y: 220 },
-    ]);
-    for (let index = 1; index < points.length; index += 1) {
-      const previous = points[index - 1]!;
-      const next = points[index]!;
-      expect(previous.x === next.x || previous.y === next.y).toBe(true);
-    }
-  });
-
   it("spreads multiple schema edges across table header lanes", async () => {
     const snapshot = {
       viewType: "inlineC4" as const,
@@ -1491,50 +1464,6 @@ describe("SoftwareMap inline C4 helpers", () => {
         ),
       ),
     ).toEqual([false, false]);
-  });
-
-  it("keeps positioned C4 edge labels on their edges while avoiding overlaps", () => {
-    const edgeSections = new Map([
-      [
-        "edge-a",
-        [
-          {
-            startPoint: { x: 0, y: 0 },
-            endPoint: { x: 420, y: 0 },
-          },
-        ],
-      ],
-      [
-        "edge-b",
-        [
-          {
-            startPoint: { x: 0, y: 0 },
-            endPoint: { x: 420, y: 0 },
-          },
-        ],
-      ],
-    ]);
-    const edgeLabels = new Map([
-      ["edge-a", { x: 180, y: -12, width: 96, height: 24 }],
-      ["edge-b", { x: 180, y: -12, width: 96, height: 24 }],
-    ]);
-    const nodeObstacles = [{ x: 150, y: -44, width: 120, height: 88 }];
-
-    const positioned = positionC4EdgeLabels(
-      edgeSections,
-      edgeLabels,
-      nodeObstacles,
-    );
-    const first = positioned.get("edge-a");
-    const second = positioned.get("edge-b");
-
-    expect(first).toBeDefined();
-    expect(second).toBeDefined();
-    expect(first!.y + first!.height / 2).toBe(0);
-    expect(second!.y + second!.height / 2).toBe(0);
-    expect(labelBoxesOverlapForTest(first!, nodeObstacles[0]!)).toBe(false);
-    expect(labelBoxesOverlapForTest(second!, nodeObstacles[0]!)).toBe(false);
-    expect(labelBoxesOverlapForTest(first!, second!)).toBe(false);
   });
 
   it("keeps node comment buttons beside world-space nodes", () => {
@@ -2873,16 +2802,4 @@ function isOrthogonalPolylineForTest(points: Array<{ x: number; y: number }>) {
     }
   }
   return true;
-}
-
-function labelBoxesOverlapForTest(
-  left: { x: number; y: number; width: number; height: number },
-  right: { x: number; y: number; width: number; height: number },
-) {
-  return !(
-    left.x + left.width <= right.x ||
-    right.x + right.width <= left.x ||
-    left.y + left.height <= right.y ||
-    right.y + right.height <= left.y
-  );
 }
